Change libraian-gc file deletion to autocommit
In pyscopg2, even "SELECT" queries are counted as transactions by
default. And currently, librarian-gc uses autocommit with the value
False. This results in a need to close each transaction manually, which
were logical while we were using a single large query, but since we
seperated the queries  into chunks, this results in each query to be
kept open. Thus we change the autocommit value to True so that after
each query we start working on the next query.