Add rlimit preservation context manager and apply
Jobs have a memory limit baked in that is set with a `setrlimit()`
call. Normally this would only be called within Celery tasks, but due
to the way we are doing tests in buildbot, the `setrlimit` gets called
within the main testing process of the buildbot worker when called
with the `runAll` method. This causes the memory's of the tests to be
limited and result in failures for certain tests due to Memory and
thread creation errors.

As a result, I added a context manager that preserves and restores
the rlimit in question whenever `runAll` is called in the tests
in question.