Add an additional memory preserver to missed test
Another test, this time inside the BranchScanJob, was silently
modifying the memory limit, as such, we wrapped it with
"preserve_limit()".

To avoid these memory changes, the preserve_limit context manager first
saves the current RLIMIT_AS resource, which denotes the virtual memory
limit, and then resets it to the saved value after the run has
completed.

Also added more descriptive explanations and reduced the scope of
the error catch within the new context manager.