Allow webhooks to be configured on archives
This is part of the implementation to add webhooks for
source-package-upload on archives. In this MP, we:
Inherit the necessary Webhook classes in Archive classes.
Add archive as a target for webhooks.
Add unit tests to ensure that webhooks can be configured on
archives and deleted once the archive is deleted.