Ensure event.edited_fields exists in source package webhook handler
This MP fixes the tests broken by
https://code.launchpad.net/~vaishnavi-asawale/launchpad/+git/launchpad/+merge/494750
in which, as a final step, the lazr.restful layer triggers the ObjectModifiedEvent
with a security-proxied PackageUpload object that has edited_fields set to None in
the modified event, causing the test to break at the handler function which treats
it as an iterable and checks for the existence of 'status' in the edited_fields.

This is fixed by first ensuring that event.edited_fields is not None.